+++ packages/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @render-harness/web
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @render-harness/registry@0.2.2
+
 ## 0.2.2
 
 ### Patch Changes